requirements.txt
openai>=1.40.0
python-dotenv>=1.0.1
Flask>=3.0.0
httpx>=0.27.0
